Fix #238: don't find statements twice!
Racer does two passes when searching a scope, one backwards from point and one forwards (so that it can track let expressions in the right order). It uses a rust statement iterator to do the passes, but I had marked StmtIndicesIter 'Copy', which meant in the for-loops racer was getting a fresh copied statement iterator each time, causing it to iterate the statements twice.
The fix was to use a mutable reference for the loops rather than using the iterator by value (and implicitly copying it).