Make it possible to complete on temporary file
Most editors use a temporary file over the current file to send a
completion or find-definition request to racer. If this temporary
file is not in the project directory, the completion will not work
for external crates as racer searches a Cargo file in the parent
directories of the file.

To fix this problem and to not break backward compatibility this commit
adds an optional argument to specify the file name which belongs to the
project.