Automatic completion on certain prefixes, annotation improvements, location support.
The completion now begins automatically (customizable) if point is preceded by :: or .

Annotations now appear on the first completion (there was some weird thing that caused
them to only appear after some interaction had occurred).

The company eldoc strings are now syntax hilighted.

Location jumping: hitting C-w on a candidate takes you to the file/line of the match.
Column jump should be possible but company does not support that.

`company-idle-delay` setting to nil was removed: the user should decide
if the triggering is manual or not.