Wrapped complete in a thread to catch panics
Plugins are calling racer on every keypress, and then leaving tmp files
around on non-zero exits. For now I am wrapping 'complete' in a thread
to catch panics and exit nicely

This probably sucks performance-wise, sorry!

Also passed match_fn as an enum because &Fn() trait is not sync. There
is probably a better way to do this but figured passing an enum is no
biggie