Session supports preloaded files
Session.load_file is used during calls to find_definition to get the
querying file's contents. The load_file method reads the file from disk.
This presents issues for new files and files with unsaved changes that
are being queried from racer. The issue arises when using racer as a
library and the embedding application has access to the unsaved buffer.

The method `cache_file_contents` was added to Session to support that
use case. The first argument is a path as before, and the second
argument is any type implementing `Into<String>` which is already
implemented by &str and String (no-op in this case).

By precaching the file, subsequent calls to `load_file` will return
the cached contents.