Fix Offsets of CRLF files
coords_to_point and point_to_coords_from_file assume that the line endings
are ending with a simple LF and not a CRLF, which is why they get the
offset wrong which then causes Racer to complete the wrong parts of the
code. By using split('\n') instead of lines(), the code can still assume
that the LF is missing, but the CR is still part of the individual splits,
so it gets counted properly.

Fixes #435