Fix FileCache safety
The `increment_generation` function was inherently unsafe in the past,
but it was not declared as part of the API. Thus, consumers could have
called it without any warning from the compiler.

The allocation management has changed as well. The increment_generation
function now converts freed references into mutable references. Dropping
is still delayed until an allocation is reused since otherwise a double
free could happen when the Arena is deallocated.

The overwriting_cached_files test now tests the allocation reuse code
paths.