Fix formatting of contextstr for refs/vecs
Given the following example,

    struct Foo<'a> {
        reference: &'a u8,
        array: [u8; 5],
        slice: &'a [u8],
    }

    fn foo(x: Foo) {
        x.
    }

The contextstr for completions should be &u8, [u8; 5], and &[u8] for
each Foo member, respectively. This patch fixes behavior where
contextstr was u8, array, and slice instead.

Variants were added to core::Ty to represent ref ptrs, fixed length
vecs, and vecs. The Display impl for Ty was updated to format them as
expected.

Perhaps questionably, these new variants hold a Box for another Ty. This
makes awkward anywhere that needs to destructure the variant and pass
the nested type as an owned object.