Fix #470
I have added a configuration section to the README.md which includes the
information that is common to all installation methods (downloading the
rust src). I have also added a step to add the binary to your `PATH`
from the source installation method. I have added a section about the
racer atom package as well.