Fix issue with Cargo.lock crate lookup
The main loop of `racer::cargo::get_cargo_packages` previously used
otry! extensively which would cause the entire function to return None
in the case that any of the crates in the package list didn't have all
of the desired properties. `otry!` usage in this function was replaced
with an `unwrap_or_continue!` macro which yields the value of some or
`continue`s which skips the current package.