Remove test_fixtures dependency
This was preventing us from publishing to crates.io.

The test_fixtures directory was moved into a test project. The tests
that use those fixtures now change directory during the test. This is a
problem for other tests which expect to be in the normal directory or at
least not change mid test. To deal with that, and to prevent developers
from needing to run tests with RUST_TEST_THREADS=1, each test takes a
shared lock to synchronize them. In an ideal world, the test threads
flag would just be an option in Cargo.toml, or something like
https://github.com/rust-lang/rust/issues/33519 would exist.