Fix #683: do not ignore empty module doc lines
Fix for phildawes/racer/#683: racer expected an empty documentation line to start with `//! `, including a space. The rust documentation does not include this space, thus racer ignores all empty lines of the documentation, resulting in markdown being rendered differently in some cases (see referenced issue). Modify the pattern to `//!`, excluding the space, to fix this.