Delay all imports
An import may bring a name into the value namespace while using a name
from the type namespace as the first segment of its path. For example,
`use foo::foo;`, where `foo` is a module and `foo::foo` is a function.

Process imports after all other items that can introduce names in the
type namespace, so that an import doesn't try to use itself to resolve
its path, causing unbound recursion.

A test was added for this case, and it verifies that both the module and
the function are suggested as completions when relevant.

Note that glob imports are still processed after single imports, not
only for performance reasons, but because single imports can shadow
names introduced by a glob import. A test was added to verify this.

Fixes #354 (partially).