Don't overflow the stack when trying to resolve unresolvable imports
When the first and the last component of an import's path are the same,
and that name is undefined (i.e. the compiler would fail to resolve the
import), we would try to resolve the name recursively endlessly.

This commit introduces a stack of imports pending resolution. When a
`use` item is encountered, if it's already on the stack, it will be
ignored.

Fixes #354.