More helpful error when rust-src is missing
Note that this changes the behavior when "rustc --print sysroot" returns a
value. Before the change, if this command was successful, but the expected paths
not found in the yielded root, it would fallback on looking for the dir in
"/usr/...". I think this makes sense, if rustc is present and functional, it
should use its sysroot or fail. Otherwise, things are potentially inconsistent.

Also makes the following changes:

* Fixes environment variable name in description()

* Fixes double use of " around paths, display for PathBuf already includes
  quotes.