Fix panic on complete before first block statement  (#823)
* Add completion tests

Completions produce panics in cases that are reflected by the new tests.

* Fix panic on complete between statements

`find_stmt_start` would find a statement that ends after the current
point and `complete_from_file_` would then assume that the resulting
statement is the current one and will take a slice of the source from the
beginning of the statement until the given point. The assumption may not
be true when the point is not in any statement, in which case
`find_stmt_start` will return a statement that follows the point rather
than one that surrounds it as this will be the first one that ends after
the given point. Such statement would begin *after* the given point as
well, which would make it impossible to take a slice (slice end must be
greater than its start).

This commit fixes the issue by making sure that `find_stmt_start` only ever
returns result if it finds statements that the current position belongs to.